refactor(service): inline Google Sheets helpers + rename readIntoQueueWithState
Move withRowKey, compactGoogleSheetsMessages, addRowNumbers, and
augmentGoogleSheetsCatalog out of _shared.ts and into the single activity
that uses each, eliminating the @stripe/sync-destination-google-sheets
coupling from the generic shared file.

Rename readIntoQueueWithState → readGoogleSheetsIntoQueue (and its factory +
file) so the activity name reflects its Google Sheets-only scope.

Committed-By-Agent: claude
