+++ apps/service/src/temporal/activities/_shared.ts
@@ -1,12 +1,7 @@
 import { heartbeat } from '@temporalio/activity'
-import type { ConfiguredCatalog, Message, RecordMessage } from '@stripe/sync-engine'
+import type { Message } from '@stripe/sync-engine'
 import { Kafka } from 'kafkajs'
 import type { Producer } from 'kafkajs'
-import {
-  ROW_KEY_FIELD,
-  ROW_NUMBER_FIELD,
-  serializeRowKey,
-} from '@stripe/sync-destination-google-sheets'
 
 export interface ActivitiesContext {
   engineUrl: string
@@ -128,114 +123,6 @@ export function collectError(message: Record<string, unknown>): RunResult['error
     }
   }
   return null
-}
-
-export function withRowKey(record: RecordMessage, catalog?: ConfiguredCatalog): RecordMessage {
-  const primaryKey = catalog?.streams.find((stream) => stream.stream.name === record.record.stream)
-    ?.stream.primary_key
-  if (!primaryKey) return record
-  return {
-    ...record,
-    record: {
-      ...record.record,
-      data: {
-        ...record.record.data,
-        [ROW_KEY_FIELD]: serializeRowKey(primaryKey, record.record.data),
-      },
-    },
-  }
-}
-
-export function compactGoogleSheetsMessages(messages: Message[]): Message[] {
-  const compacted: Message[] = []
-  let pendingOrder: string[] = []
-  let pending = new Map<string, RecordMessage>()
-
-  const flushPending = () => {
-    for (const key of pendingOrder) {
-      const message = pending.get(key)
-      if (message) compacted.push(message)
-    }
-    pendingOrder = []
-    pending = new Map()
-  }
-
-  for (const message of messages) {
-    if (message.type === 'record') {
-      const rowKey =
-        typeof message.record.data[ROW_KEY_FIELD] === 'string'
-          ? message.record.data[ROW_KEY_FIELD]
-          : undefined
-      if (!rowKey) {
-        compacted.push(message)
-        continue
-      }
-      const dedupeKey = `${message.record.stream}:${rowKey}`
-      if (!pending.has(dedupeKey)) pendingOrder.push(dedupeKey)
-      pending.set(dedupeKey, message)
-      continue
-    }
-
-    if (message.type === 'state') {
-      flushPending()
-      compacted.push(message)
-    }
-  }
-
-  flushPending()
-  return compacted
-}
-
-export function addRowNumbers(
-  messages: Message[],
-  rowIndex: Record<string, Record<string, number>>
-): Message[] {
-  return messages.map((message) => {
-    if (message.type !== 'record') return message
-    const rowKey =
-      typeof message.record.data[ROW_KEY_FIELD] === 'string'
-        ? message.record.data[ROW_KEY_FIELD]
-        : undefined
-    const rowNumber = rowKey ? rowIndex[message.record.stream]?.[rowKey] : undefined
-    if (rowNumber === undefined) return message
-    return {
-      ...message,
-      record: {
-        ...message.record,
-        data: {
-          ...message.record.data,
-          [ROW_NUMBER_FIELD]: rowNumber,
-        },
-      },
-    }
-  })
-}
-
-export function augmentGoogleSheetsCatalog(catalog: ConfiguredCatalog): ConfiguredCatalog {
-  return {
-    streams: catalog.streams.map((configuredStream) => {
-      const props = configuredStream.stream.json_schema?.properties as
-        | Record<string, unknown>
-        | undefined
-
-      if (!props) return configuredStream
-
-      return {
-        ...configuredStream,
-        stream: {
-          ...configuredStream.stream,
-          json_schema: {
-            ...configuredStream.stream.json_schema,
-            properties: {
-              ...props,
-              [ROW_KEY_FIELD]: { type: 'string' },
-              [ROW_NUMBER_FIELD]: { type: 'number' },
-            },
-          },
-        },
-      }
-    }),
-  }
 }
 
 export async function drainMessages(stream: AsyncIterable<Record<string, unknown>>): Promise<{
